requirements.txt
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
idna==2.10
numpy==1.19.4
Pillow==8.0.1
pony==0.7.13
psycopg2==2.8.6
requests==2.24.0
six==1.15.0
urllib3==1.25.11
vk-api==11.9.0
